+++ .storybook/config.js
@@ -80,6 +80,7 @@ configure(() => {
   require('../examples/bpk-component-drawer/stories');
   require('../examples/bpk-component-fieldset/stories');
   require('../examples/bpk-component-flare/stories');
+  require('../examples/bpk-component-floating-notification/stories');
   require('../examples/bpk-component-form-validation/stories');
   require('../examples/bpk-component-graphic-promotion/stories');
   require('../examples/bpk-component-grid-toggle/stories');
@@ -104,6 +105,7 @@ configure(() => {
   require('../examples/bpk-component-panel/stories');
   require('../examples/bpk-component-phone-input/stories');
   require('../examples/bpk-component-popover/stories');
+  require('../examples/bpk-component-price/stories');
   require('../examples/bpk-component-progress/stories');
   require('../examples/bpk-component-radio/stories');
   require('../examples/bpk-component-rating/stories');