+++ packages/bpk-component-floating-notification/src/BpkFloatingNotification-test.js
@@ -1,0 +1,107 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2016 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+/* @flow strict */
+
+import React from 'react';
+import { fireEvent, render } from '@testing-library/react';
+import { cssModules } from 'bpk-react-utils';
+
+import BpkIconHeart from '../../bpk-component-icon/sm/heart';
+
+import BpkFloatingNotification, { TYPE } from './BpkFloatingNotification';
+import STYLES from './BpkFloatingNotification.module.scss';
+
+const getClassName = cssModules(STYLES);
+
+const props = {
+  text: 'Saved',
+};
+
+describe('BpkFloatingNotification', () => {
+  it('should render correctly', () => {
+    const { asFragment } = render(<BpkFloatingNotification {...props} />);
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should support custom class names', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification className="custom-classname" {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should support arbitrary props', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification testid="123" {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should render correctly in dark mode (prop)', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification type={TYPE.dark} {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should render correctly with icon prop', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification icon={BpkIconHeart} {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should render correctly with icon prop in dark mode', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification
+        type={TYPE.dark}
+        icon={BpkIconHeart}
+        {...props}
+      />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should render correctly with CTA text', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification ctaText="View" {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should render correctly with CTA text in dark mode', () => {
+    const { asFragment } = render(
+      <BpkFloatingNotification ctaText="View" type={TYPE.dark} {...props} />,
+    );
+    expect(asFragment()).toMatchSnapshot();
+  });
+
+  it('should do function when button is pressed', () => {
+    const onClick = jest.fn();
+    render(
+      <BpkFloatingNotification ctaText="View" onClick={onClick} {...props} />,
+    );
+
+    const cta = document.getElementsByClassName(
+      getClassName('bpk-floating-notification__button'),
+    )[0];
+    fireEvent.click(cta);
+
+    expect(onClick).toHaveBeenCalledTimes(1);
+  });
+});